# Add concurrency safe iterator wrapper ChannelLike