--- a/src/schedulers.jl
+++ b/src/schedulers.jl
@@ -233,8 +233,9 @@ end
 """
     GreedyScheduler (aka :greedy)
 
-A greedy dynamic scheduler. The elements of the collection are first put into a `Channel`
-and then dynamic, non-sticky tasks are spawned to process the channel content in parallel.
+A greedy dynamic scheduler. The elements are put into a shared workqueue and dynamic,
+non-sticky, tasks are spawned to process the elements of the queue with each task taking a new
+element from the queue as soon as the previous one is done.
 
 Note that elements are processed in a non-deterministic order, and thus a potential reducing
 function **must** be [commutative](https://en.wikipedia.org/wiki/Commutative_property) in
@@ -249,10 +250,10 @@ some additional overhead.
     * Determines the number of parallel tasks to be spawned.
     * Setting `ntasks < nthreads()` is an effective way to use only a subset of the available threads.
 - `chunking::Bool` (default `false`):
-    * Controls whether input elements are grouped into chunks (`true`) or not (`false`) before put into the channel. This can improve the performance especially if there are many iterations each of which are computationally cheap.
+    * Controls whether input elements are grouped into chunks (`true`) or not (`false`) before put into the shared workqueue. This can improve the performance especially if there are many iterations each of which are computationally cheap.
     * If `nchunks` or `chunksize` are explicitly specified, `chunking` will be automatically set to `true`.
 - `nchunks::Integer` (default `10 * nthreads()`):
-    * Determines the number of chunks (that will eventually be put into the channel).
+    * Determines the number of chunks (that will eventually be put into the shared workqueue).
     * Increasing `nchunks` can help with [load balancing](https://en.wikipedia.org/wiki/Load_balancing_(computing)). For `nchunks <= nthreads()` there are not enough chunks for any load balancing.
 - `chunksize::Integer` (default not set)
     * Specifies the desired chunk size (instead of the number of chunks).