Just to show that the idea from https://github.com/JuliaFolds2/ChunkSplitters.jl/issues/49 can be used to improve the (chunked) GreedyScheduler.
The advantages are that the iterator doesn't have to be copied into a Channel, and that the AtomicChunk-thingy is more efficient (but not as general, of course) than a Channel. See some benchmarks in the test.jl file.