@MasonProtter 2024-09-25
Very nice idea, I had wanted something like this before.