@carstenbauer 2024-09-27
BTW, could perhapse make sense to mention this in the docs under https://juliafolds2.github.io/OhMyThreads.jl/stable/literate/tls/tls/#Another-safe-way-based-on-Channel.