@fredrikekre 2024-09-27
Added to changelog and updated the GreedyScheduler docstring.