@carstenbauer 2024-09-27
Unfortunately, we mention "channel" more times in the docstring in the keyword arguments section. Would be great if you could "fix" this.