@fredrikekre 2024-09-27
> BTW, could perhapse make sense to mention this in the docs under

Done.

> Unfortunately, we mention "channel" more times in the docstring

Fixed, replaced with "shared workqueue".